refactor: move profile and AI name fetching to PreferencesContext
- Added avatarUrl, fullName, and aiName to PreferencesContext
- Implemented fetchUserData logic in PreferencesProvider's useEffect
- Refactored AppShell to consume these values from usePreferences()
- Removed local state and redundant fetching from AppShell
- Derived profileLoaded from fullName in AppShell
